+++ R/tar_quarto_rep_raw.R
@@ -118,6 +118,7 @@ tar_quarto_rep_raw <- function(
   cache_refresh = FALSE,
   debug = FALSE,
   quiet = TRUE,
+  quarto_args = NULL,
   pandoc_args = NULL,
   rep_workers = 1,
   packages = targets::tar_option_get("packages"),
@@ -158,6 +159,7 @@ tar_quarto_rep_raw <- function(
   targets::tar_assert_lgl(debug)
   targets::tar_assert_scalar(quiet)
   targets::tar_assert_lgl(quiet)
+  targets::tar_assert_chr(quarto_args %|||% ".")
   targets::tar_assert_chr(pandoc_args %|||% ".")
   tar_assert_rep_workers(rep_workers)
   rep_workers <- as.integer(rep_workers)
@@ -199,6 +201,7 @@ tar_quarto_rep_raw <- function(
       cache_refresh = cache_refresh,
       debug = debug,
       quiet = quiet,
+      quarto_args = quarto_args,
       pandoc_args = pandoc_args,
       default_output_file = default_output_file,
       rep_workers = rep_workers
@@ -309,6 +312,7 @@ tar_quarto_rep_command <- function(
   cache_refresh,
   debug,
   quiet,
+  quarto_args,
   pandoc_args,
   default_output_file,
   rep_workers
@@ -325,6 +329,7 @@ tar_quarto_rep_command <- function(
       cache_refresh = cache_refresh,
       debug = debug,
       quiet = quiet,
+      quarto_args = quarto_args,
       pandoc_args = pandoc_args,
       as_job = FALSE
     ),
@@ -336,6 +341,7 @@ tar_quarto_rep_command <- function(
       cache_refresh = cache_refresh,
       debug = debug,
       quiet = quiet,
+      quarto_args = quarto_args,
       pandoc_args = pandoc_args
     )
   )